this may be final commit, deleted one line
this is really hard to test. there is no way to know that their email actually exists. but they can only login if their email has been verified so Im not sure it matters. maybe autodelete accounts that have not been verified for a period of time so the database doesn't overpopulate